API Documentation

CIF files

A Cif is a collection of CifContainers indexed by a String label.

CrystalInfoFramework.Cif — Method

Cif(s::AbstractString;verbose=false)

Read in filename s as a CIF file. If verbose is true, print progress information during parsing.

Base.keys — Method

keys(c::Cif)

The names of all blocks in c, not including any save frames.

Base.first — Method

first(c::Cif)

The first block in c, which may not be the first block that appears in the physical file. This is useful when only one block is present.

Base.length — Method

length(c::Cif)

The number of blocks in n.

Base.haskey — Method

haskey(c::Cif,name)

Whether c has a block named name.

Base.getindex — Method

getindex(c::Cif,n)

c[n] returns the block named n in c.

Base.setindex! — Method

setindex!(c::Cif,v,n)

c[n] = s sets block n to v in c.

Base.show — Method

show(io::IO,::MIME"text/plain",c::Cif)

Display a text representation of c to io. This text representation is not guaranteed to be syntactically correct CIF. To display c as a CIF file, use ::MIME"text/cif".

CIF blocks

Concrete types of CifContainers are Block and CifBlock. Only the latter may contain nested save frames. CifContainers act like Dict{String,Array{CifValue,1}} dictionaries indexed by data name.

CrystalInfoFramework.Block — Type

A CIF data block or save frame containing no nested save frames.

CrystalInfoFramework.CifBlock — Type

A CIF block potentially containing save frames.

Base.keys — Method

keys(b::CifContainer)

All data names in b

Base.haskey — Method

haskey(b::CifContainer,s::String)

Returns true if b contains a value for case-insensitive data name s

Base.iterate — Method

iterate(b::CifContainer)

Iterate over all data names in b.

Base.getindex — Method

getindex(b::CifContainer,s::String)

b[s] returns all values for case-insensitive data name s in b as an Array{CifValue,1}

Base.get — Method

get(b::CifContainer,s::String,a)

Return b[s]. If s is missing, return a. s is case-insensitive.

Base.getindex — Method

getindex(b::CifContainer,s::Dict)

Return the set of values in b corresponding to the data name values provided in s. The keys of s must be datanames found in b. A DataFrame is returned. The keys of s are case-insensitive.

Base.setindex! — Method

setindex!(b::CifContainer,v,s)

Set the value of s in b to v

Base.delete! — Method

delete!(b::CifContainer,s)

Remove the value of s from b

CIF values

Data names are associated with arrays of values of type CifValue. Single-valued data names are associated with arrays with a single element.

CrystalInfoFramework.CifValue — Type

The syntactical type of data held in a CIF file. A value is of type String, Vector{CifValue}, Dict{String,CifValue}, Missing or Nothing. In all cases the values returned for a given data name are in an Array{CifValue,1}.

Loops

CIF blocks contain key-value pairs and loops. The following methods give access to these loops.

CrystalInfoFramework.get_loop — Function

get_loop(b::CifContainer,s)

A DataFrame built from data items in the same loop as s. If no data are available, an empty DataFrame is returned.

CrystalInfoFramework.add_to_loop! — Function

add_to_loop!(b::CifContainer, tgt, newname)

Add dataname tgt to the loop containing newname. Values for tgt must already be present (e.g. by calling b[tgt]=values) and have the same length as other values in the loop.

CrystalInfoFramework.create_loop! — Function

create_loop!(b::CifContainer,names::Array{String,1})

Create a loop in b from the datanames in names. Datanames previously assigned to other loops are transferred to the new loop. All data attached to names should have the same length.

Save frames

CIF data files do not contain save frames, however CIF dictionaries use them extensively. The contents of save frames are invisible to all methods except show. They can be accessed using get_frames, which returns a Cif object.

CrystalInfoFramework.get_frames — Method

get_frames(f::CifBlock{V})

Return all nested CIF containers in f as a Cif object.