API Documentation

CIF files

A Cif is a collection of CifContainers indexed by a String label.

CrystalInfoFramework.Cif — Method

Cif(s::AbstractString;verbose=false)

Read in filename s as a CIF file. If verbose is true, print progress information during parsing.

Base.keys — Method

keys(c::Cif)

The names of all blocks in c, not including any save frames.

Base.first — Method

first(c::Cif)

The first block in c, which may not be the first block that appears in the physical file. This is useful when only one block is present.

Base.length — Method

length(c::Cif)

The number of blocks in n.

Base.haskey — Method

haskey(c::Cif,name)

Whether c has a block named name.

Base.getindex — Method

getindex(c::Cif,n)

c[n] returns the block named n in c.

Base.setindex! — Method

setindex!(c::Cif,v,n)

c[n] = s sets block n to v in c.

Base.show — Method

show(io::IO,::MIME"text/plain",c::Cif)

Display a text representation of c to io. This text representation is not guaranteed to be syntactically correct CIF. To display c as a CIF file, use ::MIME"text/cif".

CIF blocks

Concrete types of CifContainers are Block and CifBlock. Only the latter may contain nested save frames. CifContainers act like Dict{String,Array{CifValue,1}} dictionaries indexed by data name.

CrystalInfoFramework.Block — Type

A CIF data block or save frame containing no nested save frames.

CrystalInfoFramework.CifBlock — Type

A CIF block potentially containing save frames.

Base.keys — Method

keys(b::CifContainer)

All data names in b

Base.haskey — Method

haskey(b::CifContainer,s::String)

Returns true if b contains a value for case-insensitive data name s

Base.iterate — Method

iterate(b::CifContainer)

Iterate over all data names in b.

Base.getindex — Method

getindex(b::CifContainer,s::String)

b[s] returns all values for case-insensitive data name s in b as an Array{CifValue,1}

Base.get — Method

get(b::CifContainer,s::String,a)

Return b[s]. If s is missing, return a. s is case-insensitive.

Base.getindex — Method

getindex(b::CifContainer,s::Dict)

Return the set of values in b corresponding to the data name values provided in s. The keys of s must be datanames found in b. A DataFrame is returned. The keys of s are case-insensitive.

Base.setindex! — Method

setindex!(b::CifContainer,v,s)

Set the value of s in b to v

Base.delete! — Method

delete!(b::CifContainer,s)

Remove the value of s from b

CIF values

Data names are associated with arrays of values of type CifValue. Single-valued data names are associated with arrays with a single element.

CrystalInfoFramework.CifValue — Type

The syntactical type of data held in a CIF file. A value is of type String, Vector{CifValue}, Dict{String,CifValue}, Missing or Nothing. In all cases the values returned for a given data name are in an Array{CifValue,1}.

Loops

CIF blocks contain key-value pairs and loops. The following methods give access to these loops.

CrystalInfoFramework.get_loop — Function

get_loop(b::CifContainer,s)

A DataFrame built from data items in the same loop as s. If no data are available, an empty DataFrame is returned.

CrystalInfoFramework.add_to_loop! — Function

add_to_loop!(b::CifContainer, tgt, newname)

Add dataname tgt to the loop containing newname. Values for tgt must already be present (e.g. by calling b[tgt]=values) and have the same length as other values in the loop.

CrystalInfoFramework.create_loop! — Function

create_loop!(b::CifContainer,names::Array{String,1})

Create a loop in b from the datanames in names. Datanames previously assigned to other loops are transferred to the new loop. All data attached to names should have the same length.

Save frames

CIF data files do not contain save frames, however CIF dictionaries use them extensively. The contents of save frames are invisible to all methods except show. They can be accessed using get_frames, which returns a Cif object.

CrystalInfoFramework.get_frames — Method

get_frames(f::CifBlock{V})

Return all nested CIF containers in f as a Cif object.

CIF dictionaries

DDLm Dictionaries

Dictionaries published by the International Union of Crystallography use the DDLm attribute set.

CrystalInfoFramework.DDLm_Dictionary — Type

A DDLm Dictionary holds information about data names including executable methods for deriving missing values.

Missing docstring.

Missing docstring for DDLm_Dictionary(c::Cif). Check Documenter's build log for details.

Missing docstring.

Missing docstring for DDLm_Dictionary(a::String;verbose=false). Check Documenter's build log for details.

Missing docstring.

Missing docstring for keys(d::DDLm_Dictionary). Check Documenter's build log for details.

Missing docstring.

Missing docstring for getindex(d::DDLm_Dictionary,k). Check Documenter's build log for details.

Base.delete! — Method

delete!(d::DDLm_Dictionary,k::String)

Remove all information associated with dataname k

Missing docstring.

Missing docstring for get_dic_namespace(d::DDLm_Dictionary). Check Documenter's build log for details.

CrystalInfoFramework.list_aliases — Method

list_aliases(d::DDLm_Dictionary,name;include_self=false)

List aliases of name listed in d. If not include_self, remove name from the returned list.

CrystalInfoFramework.find_name — Method

find_name(d::DDLm_Dictionary,name)

Find the canonical name for name in d. If name is not present, return name unchanged. If accessed in cat/obj format, search also child categories. Note that the head category may not have a category associated with it.

CrystalInfoFramework.find_category — Method

find_category(d::DDLm_Dictionary,dataname)

Find the category of dataname by looking up d.

CrystalInfoFramework.find_object — Method

find_object(d::DDLm_Dictionary,dataname)

Find the object_id of dataname by looking up d.

CrystalInfoFramework.is_category — Method

is_category(d::DDLm_Dictionary,name)

Return true if name is a category according to d.

CrystalInfoFramework.get_categories — Method

get_categories(d)

List all categories defined in DDLm Dictionary d

CrystalInfoFramework.get_keys_for_cat — Method

get_keys_for_cat(d::DDLm_Dictionary,cat;aliases=false)

List all category key data names for cat listed in d. If aliases, include alternative names for the key data names.

CrystalInfoFramework.get_default — Method

get_default(d::DDLm_Dictionary,s)

Return the default value for s or missing if none defined.

CrystalInfoFramework.lookup_default — Method

lookup_default(dict::DDLm_Dictionary,dataname::String,cp)

Index into any default lookup table defined in dict for dataname using an index value from cp. cp must have a property name as specified by def_index_id in the definition of dataname such that cp.<def_index_id> returns a single value

Base.show — Method

show(io::IO,::MIME"text/cif",ddlm_dic::DDLm_Dictionary)

Output ddlm_dic in CIF format.

DDL2 Dictionaries

DDL2 dictionaries are published and maintained by the worldwide Protein Data Bank (wwPDB).

CrystalInfoFramework.DDL2_Dictionary — Type

DDL2Dictionary(ds,attdic::DDL2_Dictionary,dividers)

Create a DDL2_Dictionary from ds, using the category scheme and attributes in att_dic, sorting definitions based on the attributes in dividers. ds must contain _dictionary.title

Construct a dictionary when provided with a collection of data frames indexed by symbols. The symbols are DDL2 attribute categories, and the dataframe columns are the object_ids of the DDL2 attributes of that category.

TODO: work out child-parent relations as well

Missing docstring.

Missing docstring for DDL2_Dictionary(c::Cif). Check Documenter's build log for details.

Missing docstring.

Missing docstring for DDL2_Dictionary(a::String;verbose=false). Check Documenter's build log for details.

Missing docstring.

Missing docstring for keys(d::DDL2_Dictionary). Check Documenter's build log for details.

Missing docstring.

Missing docstring for getindex(d::DDL2_Dictionary,k). Check Documenter's build log for details.

CrystalInfoFramework.get_categories — Method

get_categories(d)

List all categories defined in DDLm Dictionary d

Missing docstring.

Missing docstring for get_default(d::DDL2_Dictionary,dataname). Check Documenter's build log for details.

Base.show — Method

show(io::IO, mime, x)

The display functions ultimately call show in order to write an object x as a given mime type to a given I/O stream io (usually a memory buffer), if possible. In order to provide a rich multimedia representation of a user-defined type T, it is only necessary to define a new show method for T, via: show(io, ::MIME"mime", x::T) = ..., where mime is a MIME-type string and the function body calls write (or similar) to write that representation of x to io. (Note that the MIME"" notation only supports literal strings; to construct MIME types in a more flexible manner use MIME{Symbol("")}.)

For example, if you define a MyImage type and know how to write it to a PNG file, you could define a function show(io, ::MIME"image/png", x::MyImage) = ... to allow your images to be displayed on any PNG-capable AbstractDisplay (such as IJulia). As usual, be sure to import Base.show in order to add new methods to the built-in Julia function show.

Technically, the MIME"mime" macro defines a singleton type for the given mime string, which allows us to exploit Julia's dispatch mechanisms in determining how to display objects of any given type.

The default MIME type is MIME"text/plain". There is a fallback definition for text/plain output that calls show with 2 arguments, so it is not always necessary to add a method for that case. If a type benefits from custom human-readable output though, show(::IO, ::MIME"text/plain", ::T) should be defined. For example, the Day type uses 1 day as the output for the text/plain MIME type, and Day(1) as the output of 2-argument show.

Container types generally implement 3-argument show by calling show(io, MIME"text/plain"(), x) for elements x, with :compact => true set in an IOContext passed as the first argument.

Data Sources

Data from arbitrary file formats can be used as long as they return an array of values when provided with a string.

CrystalInfoFramework.DataContainer.DataSource — Type

A dictionary is a data source

A Cif Block is a data source. It implements the dictionary interface.

Missing docstring.

Missing docstring for TypedDataSource. Check Documenter's build log for details.