API Documentation

CIF files

A Cif is a collection of CifContainers indexed by a String label.

CrystalInfoFramework.Cif — Method

Cif(s::AbstractPath;verbose=false,native=false,version=0)

Read in filename s as a CIF file. If verbose is true, print progress information during parsing. If native is false, use the C-language parser provided by libcif. If libcif is not available, the native Julia parser will be used. version may be 1, 2 or 0 (default) for auto-detected CIF version. version is only respected by the native parser. The libcif parser will always auto-detect.

CrystalInfoFramework.Cif — Method

Cif(s::AbstractString;verbose=false,version=0,source=p"")

Process contents of s as a CIF file using the native Julia parser. If verbose is true, print progress information during parsing. version may be 1, 2 or 0 (default) for auto-detected CIF version. If source is provided, it is a filesystem location to record as the source for s.

Base.keys — Method

keys(c::Cif)

The names of all blocks in c, not including any save frames.

Base.first — Method

first(c::Cif)

The first block in c, which may not be the first block that appears in the physical file. This is useful when only one block is present.

Base.length — Method

length(c::Cif)

The number of blocks in n.

Base.haskey — Method

haskey(c::Cif,name)

Whether c has a block named name.

Base.getindex — Method

getindex(c::Cif,n)

c[n] returns the block named n in c.

Base.setindex! — Method

setindex!(c::Cif,v,n)

c[n] = s sets block n to v in c.

Base.show — Method

show(io::IO,::MIME"text/plain",c::Cif)

Display a text representation of c to io. This text representation is not guaranteed to be syntactically correct CIF. To display c as a CIF file, use ::MIME"text/cif".

CIF blocks

Concrete types of CifContainers are Block and CifBlock. Only the latter may contain nested save frames. CifContainers act like Dict{String,Array{CifValue,1}} dictionaries indexed by data name.

CrystalInfoFramework.Block — Type

A CIF data block or save frame containing no nested save frames.

CrystalInfoFramework.CifBlock — Type

A CIF block potentially containing save frames. Save frames cannot be nested.

Base.keys — Method

keys(b::CifContainer)

All data names in b

Base.haskey — Method

haskey(b::CifContainer,s::String)

Returns true if b contains a value for case-insensitive data name s

Base.iterate — Method

iterate(b::CifContainer)

Iterate over all data names in b.

Base.getindex — Method

getindex(b::CifContainer,s::String)

b[s] returns all values for case-insensitive data name s in b as an Array{CifValue,1}

Base.get — Method

get(b::CifContainer,s::String,a)

Return b[s]. If s is missing, return a. s is case-insensitive.

Base.getindex — Method

getindex(b::CifContainer,s::Dict)

Return the set of values in b corresponding to the data name values provided in s. The keys of s must be datanames found in b. A DataFrame is returned. The keys of s are case-insensitive.

Base.setindex! — Method

setindex!(b::CifContainer,v,s)

Set the value of s in b to v

Base.delete! — Method

delete!(b::CifContainer,s)

Remove the value of s from b

CIF values

Data names are associated with arrays of values of type CifValue. Single-valued data names are associated with arrays with a single element.

CrystalInfoFramework.CifValue — Type

The syntactical type of data held in a CIF file. A value is of type String, Vector{CifValue}, Dict{String,CifValue}, Missing or Nothing. In all cases the values returned for a given data name are in an Array{CifValue,1}.

Loops

CIF blocks contain key-value pairs and loops. The following methods give access to these loops.

CrystalInfoFramework.get_loop — Function

get_loop(b::CifContainer,s)

A DataFrame built from data items in the same loop as s. If no data are available, an empty DataFrame is returned.

CrystalInfoFramework.add_to_loop! — Function

add_to_loop!(b::CifContainer, tgt, newname)

Add dataname tgt to the loop containing newname. Values for tgt must already be present (e.g. by calling b[tgt]=values) and have the same length as other values in the loop.

CrystalInfoFramework.create_loop! — Function

create_loop!(b::CifContainer,names)

Create a loop in b from the datanames in names. Datanames previously assigned to other loops are transferred to the new loop. All data attached to names should have the same length.

Save frames

CIF data files do not contain save frames, however CIF dictionaries use them extensively. The contents of save frames are invisible to all methods except show. They can be accessed using get_frames, which returns a Cif object.

CrystalInfoFramework.get_frames — Method

get_frames(f::CifBlock{V})

Return all nested CIF containers in f as a Cif object.

CIF dictionaries

DDLm Dictionaries

Dictionaries published by the International Union of Crystallography use the DDLm attribute set.

CrystalInfoFramework.DDLm_Dictionary — Type

A DDLm Dictionary holds information about data names including executable methods for deriving missing values.

CrystalInfoFramework.DDLm_Dictionary — Method

DDLm_Dictionary(c::Cif;ignore_imports=false)

Create a DDLm_Dictionary from c. ignore_imports = true will ignore any import attributes.

CrystalInfoFramework.DDLm_Dictionary — Method

DDLm_Dictionary(a::AbstractPath;verbose=false,ignore_imports=false)

Create a DDLm_Dictionary given filename a. verbose = true will print extra debugging information during reading.ignore_imports = true will ignore any import attributes.

Base.keys — Method

keys(d::DDLm_Dictionary)

Return a list of datanames defined by the dictionary, including any aliases.

Base.getindex — Method

getindex(d::DDLm_Dictionary,k)

d[k] returns the definition for data name k as a Dict{Symbol,DataFrame} where Symbol is the attribute category (e.g. :name).

Base.delete! — Method

delete!(d::DDLm_Dictionary,k::String)

Remove all information from d associated with dataname k

CrystalInfoFramework.get_dic_namespace — Method

get_dic_namespace(d::DDLm_Dictionary)

Return the namespace declared by the dictionary, or ddlm if none present.

CrystalInfoFramework.list_aliases — Method

list_aliases(d::DDLm_Dictionary,name;include_self=false)

List aliases of name listed in d. If not include_self, remove name from the returned list.

CrystalInfoFramework.find_name — Method

find_name(d::DDLm_Dictionary,name)

Find the canonical name for name in d. If name is not present, return name unchanged. If accessed in cat/obj format, search also child categories. Note that the head category may not have a category associated with it. If name is the dictionary title it is returned as is.

CrystalInfoFramework.find_category — Method

find_category(d::DDLm_Dictionary,dataname)

Find the category of dataname by looking up d.

CrystalInfoFramework.find_object — Method

find_object(d::DDLm_Dictionary,dataname)

Find the object_id of dataname by looking up d.

CrystalInfoFramework.is_category — Method

is_category(d::DDLm_Dictionary,name)

Return true if name is a category according to d.

CrystalInfoFramework.get_categories — Method

get_categories(d::DDLm_Dictionary)

List all categories defined in DDLm Dictionary d

CrystalInfoFramework.get_keys_for_cat — Method

get_keys_for_cat(d::DDLm_Dictionary,cat;aliases=false)

List all category key data names for cat listed in d. If aliases, include alternative names for the key data names.

CrystalInfoFramework.get_default — Method

get_default(d::DDLm_Dictionary,s)

Return the default value for s or missing if none defined.

CrystalInfoFramework.lookup_default — Method

lookup_default(dict::DDLm_Dictionary,dataname::String,cp)

Index into any default lookup table defined in dict for dataname using an index value from cp. cp is any object with a property name as specified by def_index_id in the definition of dataname such that cp.<def_index_id> returns a single value

Base.show — Method

show(io::IO,MIME(text/cif),ddlm_dic::DDLm_Dictionary;header="")

Output ddlm_dic in CIF format. header can contain text that will be output in a comment box at the top of the file. Lines may be no longer than 74 characters. Multiple lines are will be separated by spaces.

DDL2 Dictionaries

DDL2 dictionaries are published and maintained by the worldwide Protein Data Bank (wwPDB).

CrystalInfoFramework.DDL2_Dictionary — Type

DDL2_Dictionary

The type of DDL2 dictionaries.

CrystalInfoFramework.DDL2_Dictionary — Method

DDL2_Dictionary(c::Cif)

Create a DDL2_Dictionary from c.

CrystalInfoFramework.DDL2_Dictionary — Method

DDL2_Dictionary(a::AbstractPath;verbose=false)

Create a DDL2_Dictionary given filename a. verbose = true will print extra debugging information during reading.

Base.keys — Method

keys(d::DDL2_Dictionary)

Return a list of datanames defined by the dictionary, including any aliases.

Base.getindex — Method

getindex(d::DDL2_Dictionary,k)

d[k] returns the definition for data name k as a Dict{Symbol,DataFrame} where Symbol is the attribute category (e.g. :item_name).

CrystalInfoFramework.get_categories — Method

get_categories(d::DDL2_Dictionary)

List all categories defined in d

CrystalInfoFramework.get_default — Method

get_default(d::DDL2_Dictionary,dataname)

Return the default value for dataname or missing if none defined.

Base.show — Method

show(io::IO,::MIME"text/cif",ddl2_dic::DDL2_Dictionary)

Output ddl2_dic in CIF format.

Data Sources

Data from arbitrary file formats can be used as long as they return an array of values when provided with a string.

CrystalInfoFramework.DataContainer.DataSource — Type

A DataSource, when provided with a string referring to a data name from a namespace known to the data source, returns an array of values that are either Native (already Julia types) or Text (requiring interpretation to create Julia types).

CrystalInfoFramework.DataContainer.TypedDataSource — Type

TypedDataSource(data,dictionary)

A TypedDataSource is a DataSource that returns items of with the correct type and aliases resolved, as specified in the associated CIF dictionary.